+++ src/QueueManager.cpp
