+++ src/QueueManager.h
