+++ src/SensorManager.cpp
