+++ src/SensorManager.h
